Produces an HTML report in the current working directory to evaluate regressivity in the supplied jurisdiction

Produces an HTML report in the current working directory to evaluate regressivity in the supplied jurisdiction

make_report(ratios, jurisdiction_name, output_dir = "", diagnostic = FALSE)

Arguments

ratios
a dataframe which has been preprocessed by reformat_data

jurisdiction_name
the name of the jurisdiction being analyzed

output_dir
optional, specify a folder to save the output in

diagnostic
optional, if TRUE includes additional sections evaluating data quality

Value

produces a file named `jurisdiction_name.html` in the working directory